factor out narrative layout
