+++ spotlight-client/src/NarrativeLayout/NarrativeNavigation/index.ts
