+++ spotlight-client/src/NarrativeLayout/NarrativeNavigation/utils.ts
