+++ spotlight-client/src/SystemNarrativePage/SystemNarrativePage.tsx
@@ -15,182 +15,49 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
-import { navigate, useParams } from "@reach/router";
-import useBreakpoint from "@w11r/use-breakpoint";
 import HTMLReactParser from "html-react-parser";
-import React, { useCallback, useEffect, useRef, useState } from "react";
-import { InView } from "react-intersection-observer";
-import { useSpring } from "react-spring/web.cjs";
-import Sticker from "react-stickyfill";
-import { NAV_BAR_HEIGHT } from "../constants";
+import React from "react";
 import SystemNarrative from "../contentModels/SystemNarrative";
-import getUrlForResource from "../routerUtils/getUrlForResource";
-import normalizeRouteParams from "../routerUtils/normalizeRouteParams";
 import {
   Chevron,
-  NarrativeWrapper,
   NarrativeIntroContainer,
   NarrativeIntroCopy,
-  NarrativeNavContainer,
-  NarrativeNavStickyContainer,
   NarrativeScrollIndicator,
-  NarrativeSectionsContainer,
   NarrativeTitle,
 } from "../UiLibrary";
 import Section from "./Section";
-import NarrativeNavigation from "../NarrativeNavigation";
+import NarrativeLayout from "../NarrativeLayout";
 
 const SystemNarrativePage: React.FC<{
   narrative: SystemNarrative;
 }> = ({ narrative }) => {
-  const routeParams = useParams();
-  const sectionsContainerRef = useRef() as React.MutableRefObject<
-    HTMLDivElement
-  >;
-  const showSectionNavigation = useBreakpoint(true, ["mobile-", false]);
-
-  // automated scrolling is a special case of section visibility;
-  // this flag lets us suspend in-page navigation actions while it is in progress
-  // (it is a local variable rather than a piece of React state because
-  // the animation functions are outside the render loop and won't receive state updates)
-  let isScrolling = false;
-  const cancelAutoScroll = () => {
-    isScrolling = false;
-  };
-
-  const [activeSection, directlySetActiveSection] = useState(
-    // make sure we consume the section number in the URL, if any, on first mount
-    Number(routeParams.sectionNumber) || 1
-  );
-  // wrap the section state setter in a function that respects the flag
-  const setActiveSection = useCallback(
-    (sectionNumber: number) => {
-      if (!isScrolling) {
-        directlySetActiveSection(sectionNumber);
-      }
-    },
-    [isScrolling]
-  );
-  // keep section state in sync with URL if it changes externally (e.g. via nav link)
-  useEffect(() => {
-    directlySetActiveSection(Number(routeParams.sectionNumber) || 1);
-  }, [routeParams.sectionNumber]);
-
-  const [, setScrollSpring] = useSpring(() => ({
-    onFrame: (props: { top: number }) => {
-      if (isScrolling) window.scrollTo(0, props.top);
-    },
-    // set the flag while animation is in progress,
-    // and listen for user-initiated scrolling (which takes priority)
-    onRest: () => {
-      isScrolling = false;
-      window.removeEventListener("wheel", cancelAutoScroll);
-      window.removeEventListener("touchmove", cancelAutoScroll);
-    },
-    onStart: () => {
-      isScrolling = true;
-      window.addEventListener("wheel", cancelAutoScroll, { once: true });
-      window.addEventListener("touchmove", cancelAutoScroll, { once: true });
-    },
-    to: { top: window.pageYOffset },
-  }));
-
-  const { tenantId, narrativeTypeId } = normalizeRouteParams(routeParams);
-  // updating the active section has two key side effects:
-  // 1. smoothly scrolling to the active section
-  // 2. updating the page URL so the section can be linked to directly
-  useEffect(() => {
-    let scrollDestination;
-    // scroll to the corresponding section by calculating its offset
-    const desiredSection = sectionsContainerRef.current.querySelector(
-      `#section${activeSection}`
-    );
-    if (desiredSection) {
-      scrollDestination =
-        window.pageYOffset + desiredSection.getBoundingClientRect().top;
-    }
-
-    // in practice this should always be defined, this is just type safety
-    if (scrollDestination !== undefined) {
-      setScrollSpring({
-        to: { top: scrollDestination - NAV_BAR_HEIGHT },
-        from: { top: window.pageYOffset },
-        reset: true,
-      });
-
-      // these should always be defined on this page; more type safety
-      if (tenantId && narrativeTypeId) {
-        navigate(
-          `${getUrlForResource({
-            page: "narrative",
-            params: { tenantId, narrativeTypeId },
-          })}/${activeSection}`
-        );
-      }
-    }
-  }, [
-    activeSection,
-    narrativeTypeId,
-    sectionsContainerRef,
-    setScrollSpring,
-    tenantId,
-  ]);
-
   return (
-    <NarrativeWrapper>
-      {showSectionNavigation && (
-        <NarrativeNavContainer>
-          <Sticker>
-            <NarrativeNavStickyContainer>
-              <NarrativeNavigation
-                activeSection={activeSection}
-                narrative={narrative}
-              />
-            </NarrativeNavStickyContainer>
-          </Sticker>
-        </NarrativeNavContainer>
-      )}
-      <NarrativeSectionsContainer ref={sectionsContainerRef}>
-        <InView
-          as="div"
-          id="section1"
-          threshold={0.3}
-          onChange={(inView) => {
-            if (inView) setActiveSection(1);
-          }}
-        >
-          <NarrativeIntroContainer>
-            <NarrativeTitle>{narrative.title}</NarrativeTitle>
-            <NarrativeIntroCopy>
-              {HTMLReactParser(narrative.introduction)}
-            </NarrativeIntroCopy>
-            <NarrativeScrollIndicator>
-              <span>SCROLL</span>
-              <Chevron direction="down" faded />
-              <Chevron direction="down" />
-            </NarrativeScrollIndicator>
-          </NarrativeIntroContainer>
-        </InView>
-
-        {narrative.sections.map((section, index) => {
-          // the first viz section is "page 2"
-          const pageId = index + 2;
-          return (
-            <InView
-              as="div"
-              id={`section${pageId}`}
-              key={section.title}
-              threshold={0.3}
-              onChange={(inView) => {
-                if (inView) setActiveSection(pageId);
-              }}
-            >
-              <Section section={section} />
-            </InView>
-          );
-        })}
-      </NarrativeSectionsContainer>
-    </NarrativeWrapper>
+    <NarrativeLayout
+      sections={[
+        {
+          title: narrative.title,
+          contents: (
+            <NarrativeIntroContainer>
+              <NarrativeTitle>{narrative.title}</NarrativeTitle>
+              <NarrativeIntroCopy>
+                {HTMLReactParser(narrative.introduction)}
+              </NarrativeIntroCopy>
+              <NarrativeScrollIndicator>
+                <span>SCROLL</span>
+                <Chevron direction="down" faded />
+                <Chevron direction="down" />
+              </NarrativeScrollIndicator>
+            </NarrativeIntroContainer>
+          ),
+        },
+        ...narrative.sections.map((section) => {
+          return {
+            title: section.title,
+            contents: <Section section={section} />,
+          };
+        }),
+      ]}
+    />
   );
 };
